Don't overwrite PSD once it is created.
This allows the user to work in Photoshop and use all the features of Photoshop.
The PSD will be generated, if it doesn't exist. After that the PSD will be
reused and changes with Storyboarder's tools will be ignored.
This has the advantage, that information about layers or vector graphics are
retained.
The behaviour can be configured in the Preference dialog.